LEDsssss!!!!!

+++ src/main/java/org/WaialuaRobotics359/robot/subsystems/SwerveModule.java
@@ -15,6 +15,8 @@ import org.WaialuaRobotics359.lib.util.SwerveModuleConstants;
 import com.ctre.phoenix.ErrorCode;
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.DemandType;
+import com.ctre.phoenix.motorcontrol.SupplyCurrentLimitConfiguration;
+import com.ctre.phoenix.motorcontrol.TalonFXControlMode;
 import com.ctre.phoenix.motorcontrol.TalonFXInvertType;
 import com.ctre.phoenix.motorcontrol.can.TalonFX;
 import com.ctre.phoenix.sensors.CANCoder;
@@ -65,7 +67,7 @@ public class SwerveModule {
         setSpeed(desiredState, isOpenLoop);
     }
 
-    public void ForceAngle(int angle){
+    public void ForceAngle(double angle){
         mAngleMotor.set(ControlMode.Position, Conversions.degreesToFalcon(angle, Constants.Swerve.angleGearRatio));
     }
